fix: keep audit module Edge-Runtime-safe, fixing PR #1274's build failure
src/app/api/admin/feature-flags/route.ts runs on the Edge Runtime and
calls logAuditMutation (src/middleware/audit.ts), which imported
appendAuditLog from the @/lib/audit barrel. After the durable-persistence
change, that barrel transitively pulled in the pg-based repository/pool,
and pg needs Node core modules (fs, net, tls, path) that the Edge Runtime
doesn't provide -- breaking the production build with "Module not found:
Can't resolve 'fs'/'path'".

A dynamic import() gated behind a `typeof EdgeRuntime` check was not
sufficient: webpack still resolves dynamic imports into the edge bundle's
module graph regardless of the runtime branch, so the same build failure
persisted.

The actual fix is structural: src/lib/audit/store.ts is back to a pure
in-memory implementation with zero database dependency (Edge-safe). The
database-backed persistence (persistToDatabase, queryAuditLog) now lives
in a new Node.js-only src/lib/audit/persist.ts, which the @/lib/audit
barrel exports for all regular (non-Edge) callers. src/middleware/audit.ts
imports appendAuditLog directly from ./store instead of the barrel, so
its transitive import graph never reaches pg -- audit entries logged from
Edge routes land in the in-memory buffer only, which is an accepted
limitation of the Edge Runtime, not a bug.

Verified: `next build` compiles and generates all 71 pages successfully
(previously failed with a webpack module-resolution error). Split
store.test.ts's database-persistence tests into a new persist.test.ts
matching the new module boundary; all 31 tests across both files pass.
