+++ src/lib/audit/__tests__/store.test.ts
@@ -1,23 +1,5 @@
-import { describe, expect, it, vi, beforeEach } from 'vitest';
-
-const { insertMock, findManyMock } = vi.hoisted(() => ({
-  insertMock: vi.fn().mockResolvedValue(undefined),
-  findManyMock: vi.fn().mockResolvedValue({ entries: [], total: 0 }),
-}));
-
-vi.mock('@/lib/db/repositories/audit-log.repository', () => ({
-  insert: insertMock,
-  findMany: findManyMock,
-}));
-
-import { appendAuditLog, queryAuditLogs, queryAuditLog, getAuditStoreSnapshot } from '../store';
-
-vi.mock('@/lib/audit', () => ({
-  appendAuditLog,
-  queryAuditLogs,
-  queryAuditLog,
-  getAuditStoreSnapshot,
-}));
+import { describe, expect, it } from 'vitest';
+import { appendAuditLog, queryAuditLogs, getAuditStoreSnapshot } from '../store';
 
 const baseInput = {
   actorId: 'user_1',
@@ -31,16 +13,7 @@ const baseInput = {
   statusCode: 201,
 };
 
-function flushSetImmediate(): Promise<void> {
-  return new Promise((resolve) => setImmediate(resolve));
-}
-
-describe('audit store', () => {
-  beforeEach(() => {
-    insertMock.mockClear();
-    findManyMock.mockClear();
-  });
-
+describe('audit store (in-memory, Edge-safe)', () => {
   it('appends a log entry with correct fields', () => {
     const entry = appendAuditLog(baseInput);
     expect(entry.id).toMatch(/^audit_/);
@@ -71,40 +44,5 @@ describe('audit store', () => {
       appendAuditLog({ ...baseInput, targetId: `bulk_${i}` });
     }
     expect(getAuditStoreSnapshot().length).toBeLessThanOrEqual(50);
-  });
-
-  it('asynchronously persists every appended entry to the database', async () => {
-    const entry = appendAuditLog({ ...baseInput, targetId: 'persist_me' });
-
-    // appendAuditLog must not block on the DB write.
-    expect(insertMock).not.toHaveBeenCalled();
-
-    await flushSetImmediate();
-
-    expect(insertMock).toHaveBeenCalledWith(expect.objectContaining({ id: entry.id, targetId: 'persist_me' }));
-  });
-
-  it('logs but does not throw when the database write fails', async () => {
-    insertMock.mockRejectedValueOnce(new Error('db down'));
-
-    expect(() => appendAuditLog({ ...baseInput, targetId: 'db_fail' })).not.toThrow();
-
-    await flushSetImmediate();
-    // allow the rejected promise's .catch handler to run
-    await flushSetImmediate();
-
-    expect(insertMock).toHaveBeenCalled();
-  });
-
-  it('queryAuditLog delegates to the database-backed repository', async () => {
-    findManyMock.mockResolvedValueOnce({
-      entries: [{ ...baseInput, id: 'audit_old', timestamp: new Date().toISOString() }],
-      total: 1,
-    });
-
-    const result = await queryAuditLog({ actorId: 'user_1' });
-
-    expect(findManyMock).toHaveBeenCalledWith({ actorId: 'user_1' });
-    expect(result.total).toBe(1);
   });
 });
