+++ src/middleware/audit.ts
@@ -1,4 +1,13 @@
-import { appendAuditLog, type AuditAction } from '@/lib/audit';
+// Imports directly from ./store (not the @/lib/audit barrel) to stay
+// Edge-Runtime-safe: this middleware is reachable from Edge routes (e.g.
+// src/app/api/admin/feature-flags/route.ts), and the barrel re-exports a
+// database-backed appendAuditLog (via ./persist.ts) that pulls in `pg`,
+// which needs Node.js core modules the Edge Runtime doesn't provide.
+// Audit entries logged from here only reach the in-memory recent-entries
+// buffer, not the durable database table — an accepted limitation of
+// Edge routes, not a bug.
+import { appendAuditLog } from '@/lib/audit/store';
+import type { AuditAction } from '@/lib/audit/types';
 
 export interface AuditMutationInput {
   action: AuditAction;
